web/backend/requirements/local.txt
-r base.txt

django-debug-toolbar==1.9.1
pylint==1.7.4
pylint-django==0.7.2
isort==4.2.15
pytest
git+https://github.com/pytest-dev/pytest-django.git
pytest-cov
pytest-factoryboy
factory-boy==2.8.1